made static boolean for testing
