+++ app/src/androidTest/java/edu/ucsd/cse110/team1_personalbest/Activities/MainActivity_CountStepUITest.java
@@ -1,11 +1,6 @@
 package edu.ucsd.cse110.team1_personalbest.Activities;
 
 
-import android.app.Activity;
-import android.app.Application;
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.test.InstrumentationRegistry;
 import android.support.test.espresso.ViewInteraction;
 import android.support.test.filters.LargeTest;
 import android.support.test.rule.ActivityTestRule;
@@ -14,22 +9,15 @@ import android.support.test.runner.AndroidJUnit4;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.ViewParent;
-import android.widget.TextView;
 
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
-import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import edu.ucsd.cse110.team1_personalbest.Fitness.Factories.FitnessServiceFactory;
-import edu.ucsd.cse110.team1_personalbest.Fitness.Interfaces.FitnessObserver;
-import edu.ucsd.cse110.team1_personalbest.Fitness.Interfaces.FitnessService;
-import edu.ucsd.cse110.team1_personalbest.Login.Factories.LoginServiceFactory;
-import edu.ucsd.cse110.team1_personalbest.Login.Interfaces.LoginService;
-import edu.ucsd.cse110.team1_personalbest.MainActivityTestRule;
 import edu.ucsd.cse110.team1_personalbest.R;
 
 import static android.support.test.espresso.Espresso.onView;
@@ -38,7 +26,6 @@ import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.withId;
 import static android.support.test.espresso.matcher.ViewMatchers.withText;
-import static edu.ucsd.cse110.team1_personalbest.Activities.MainActivity_SetGoalUITest.TEST_SERVICE;
 import static org.hamcrest.Matchers.allOf;
 
 @LargeTest
@@ -47,7 +34,7 @@ public class MainActivity_CountStepUITest {
 
 
     @Rule
-    public ActivityTestRule<MainActivity> mActivityTestRule = new MainActivityTestRule(MainActivity.class);
+    public ActivityTestRule<MainActivity> mActivityTestRule = new ActivityTestRule<>(MainActivity.class);
 
     @Rule
     public GrantPermissionRule mGrantPermissionRule =
@@ -56,6 +43,10 @@ public class MainActivity_CountStepUITest {
                     "android.permission.ACCESS_COARSE_LOCATION",
                     "android.permission.INTERNET");
 
+    @BeforeClass
+    public static void setup() {
+        MainActivity.TESTMODE = true;
+    }
 
     @Test
     public void mainActivity_CountStepUITest() {
