+++ app/src/androidTest/java/edu/ucsd/cse110/team1_personalbest/Activities/MainActivity_MainUITest.java
@@ -2,10 +2,6 @@ package edu.ucsd.cse110.team1_personalbest.Activities;
 
 
 import android.app.Activity;
-import android.app.Application;
-import android.content.Intent;
-import android.os.Bundle;
-import android.support.test.InstrumentationRegistry;
 import android.support.test.espresso.ViewInteraction;
 import android.support.test.filters.LargeTest;
 import android.support.test.rule.ActivityTestRule;
@@ -18,17 +14,15 @@ import android.view.ViewParent;
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
-import org.junit.Before;
+import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import edu.ucsd.cse110.team1_personalbest.Fitness.Factories.FitnessServiceFactory;
-import edu.ucsd.cse110.team1_personalbest.Fitness.Interfaces.FitnessObserver;
 import edu.ucsd.cse110.team1_personalbest.Fitness.Interfaces.FitnessService;
 import edu.ucsd.cse110.team1_personalbest.Login.Factories.LoginServiceFactory;
 import edu.ucsd.cse110.team1_personalbest.Login.Interfaces.LoginService;
-import edu.ucsd.cse110.team1_personalbest.MainActivityTestRule;
 import edu.ucsd.cse110.team1_personalbest.R;
 
 import static android.support.test.espresso.Espresso.onView;
@@ -45,14 +39,18 @@ import static org.hamcrest.Matchers.allOf;
 public class MainActivity_MainUITest {
 
     @Rule
-    public ActivityTestRule<MainActivity> mActivityTestRule = new MainActivityTestRule(MainActivity.class);
-
+    public ActivityTestRule<MainActivity> mActivityTestRule = new ActivityTestRule<>(MainActivity.class);
     @Rule
     public GrantPermissionRule mGrantPermissionRule =
             GrantPermissionRule.grant(
                     "android.permission.ACCESS_FINE_LOCATION",
                     "android.permission.ACCESS_COARSE_LOCATION",
                     "android.permission.INTERNET");
+
+    @BeforeClass
+    public static void setup() {
+        MainActivity.TESTMODE = true;
+    }
 
     @Test
     public void mainActivity_MainUITest() {
