+++ app/src/main/java/edu/ucsd/cse110/team1_personalbest/Fitness/Adapters/TestFitnessService.java
@@ -1,0 +1,37 @@
+package edu.ucsd.cse110.team1_personalbest.Fitness.Adapters;
+
+import edu.ucsd.cse110.team1_personalbest.Fitness.Interfaces.FitnessObserver;
+import edu.ucsd.cse110.team1_personalbest.Fitness.Interfaces.FitnessService;
+
+public class TestFitnessService implements FitnessService {
+
+    @Override
+    public int getRequestCode() {
+        return 0;
+    }
+
+    @Override
+    public void setup() {
+
+    }
+
+    @Override
+    public void startListening() {
+
+    }
+
+    @Override
+    public void stopListening() {
+
+    }
+
+    @Override
+    public void removeObservers() {
+
+    }
+
+    @Override
+    public void registerObserver(FitnessObserver observer) {
+
+    }
+}
